UART/requirements.txt
pandas
seaborn
pyserial
